Convert brookjs to typescript
The tests pass, the eslint runs & passes, but we don't have coverage.
This also doesn't work on all the other packages, so we'll convert each
of those individually. We also have to update the way we run mocha so
we glob tests written in typescript.
