--- packages/brookjs/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    "brookjs",
-    "@babel/env"
-  ],
-  "env": {
-    "test": {
-      "plugins": [
-        "@babel/transform-modules-commonjs"
-      ]
-    }
-  }
-}
