--- packages/brookjs/src/fromReduxStore.js
@@ -1,6 +1,0 @@
-import Kefir from 'kefir';
-
-export default store =>
-    Kefir.fromESObservable(store)
-        .toProperty(store.getState)
-        .setName('fromReduxStore');
