--- packages/brookjs/src/index.js
@@ -1,8 +1,0 @@
-import { eddy, loop, combineReducers } from './eddy';
-import { raf$, RAF } from './rAF';
-import fromReduxStore from './fromReduxStore';
-import ofType from './ofType';
-import observeDelta from './observeDelta';
-
-export { observeDelta, ofType, RAF, fromReduxStore, raf$, eddy, loop,
-    combineReducers };
